Update README.md
I forked this project just to have it in my github. As my mandarin is not existent, i translated just the readme.md with Deepl. Every other document, page, code is untouched. I´m not a developer. I have no clue how things are working here on GitHub and therefore if it wasnt already clear this isn´t my project and i cant guarantee that everything is correctly translated. I would appreciate if the professionals could have a look tell me whats needed to be changed.